Fixed handling of non-ascii characters
First problem was on input side, where I managed to save the
wiki text with an encoding other than utf-8. Now WIKI.txt is utf-8.

Script now assumes input is utf-8 and writes output as utf-8.
It has a bit of compatibility code to be able to do this without
worrying about whether python2 or python3 is in use.