Finish implementing "Still in the Running" section
That section happened to not be present in the old test input file,
so add another test file that includes "Still in the Running" in
addition to adding the missing pieces to the code.

The new test file also has some screenshots missing, so it generates
some valid warnings about not having them -- that's not a bug.